+++ src/ResourceManager/EventGrid/Commands.EventGrid/Properties/AssemblyInfo.cs
@@ -12,5 +12,5 @@ using System.Runtime.InteropServices;
 
 [assembly: ComVisible(false)]
 
-[assembly: AssemblyVersion("0.3.5")]
-[assembly: AssemblyFileVersion("0.3.5")]
+[assembly: AssemblyVersion("0.3.6")]
+[assembly: AssemblyFileVersion("0.3.6")]