+++ src/ResourceManager/Aks/Commands.Aks/ChangeLog.md
@@ -18,7 +18,9 @@
         - Additional information about change #1
 -->
 ## Current Release
-* Updated help files to include full parameter types.
+
+## Version 0.0.4
+* Updated help files to include full parameter types and correct input/output types.
 
 ## Version 0.0.3
 * No version update. Extracted AKS sdk to common for reuse in DevSpaces.