+++ src/ResourceManager/Relay/Commands.Relay/ChangeLog.md
@@ -18,7 +18,9 @@
         - Additional information about change #1
 -->
 ## Current Release
-* Updated help files to include full parameter types.
+
+## Version 0.3.7
+* Updated help files to include full parameter types and correct input/output types.
 
 ## Version 0.3.6
 * Updated markdown files, fix for the parameter name issue in example