+++ src/ResourceManager/ServiceFabric/Commands.ServiceFabric/ChangeLog.md
@@ -18,7 +18,9 @@
         - Additional information about change #1
 -->
 ## Current Release
-* Updated help files to include full parameter types.
+
+## Version 0.3.9
+* Updated help files to include full parameter types and correct input/output types.
 
 ## Version 0.3.8
 * Update example for `New-AzureRmServiceFabricCluster`